fix(codec): read a value from its own offset, and the same way in both wire formats
TimetzCodec.decodeBinaryAs recovers a Timestamp's sub-second part with a
second decode of the value's first eight bytes and passed a literal 0 as the
offset. A field that does not start its buffer - which no field of a composite
does - was given a fraction built from the composite's framing bytes, with the
seconds still correct. Three more read defects came out of the oracles below.

Two oracle gaps let it through, and both are now closed.

decodeBinaryOffsetInvariant compared values across offsets only for
decode(Object.class) and decodeAsString. It now sweeps every getObject(Class)
target ReadCoercions marks OK for the OID; an unpopulated OID, which is every
container, falls back to String.class. Registry-refused targets stay out:
their refusal is class-level, so it cannot depend on the offset, and building
the SQLException per fuzz trial costs more than the check is worth. An
InputStream is drained before comparing, having identity equality only.

ReadOracle.verify stays value-blind, which is its contract, so parity over the
format axis goes above it: one value, one reader, two wires, refuse together
or return equal values. No per-cell expected-value model is needed - the
driver's other leg is the expectation - so it reaches every cell the matrix
has. It holds only where both wires carry the same value, which
wiresCarryTheSameValue checks, since a generator may hand a type a value it
cannot hold and the two encoders then disagree by design.

The binary seed corpus skipped any edge case with a null value(), which is
every temporal and geometric catalogue, so those targets started from the
empty input. A read-only case now seeds from its literal, decoded into the
class the read registry names under the java.time preferences and re-encoded.
Object.class will not do there: the preferences reach decodeText, not the
decodeTextAs(Object.class) behind Codecs.decode, so timetz would arrive as a
java.sql.Time and every offset case would collapse to one degenerate wire.

The three defects the two oracles found:

PgSQLInputBinary.decodeFloat cast the value to float directly instead of
calling PrimitiveDecoders.asFloat, so readFloat kept saturating to +-Infinity
after the getters started refusing with 22003.

ByteaCodec had no decodeAsString(CharSequence), so the TextCodec default
rendered the decoded byte[] with toString() and readString on a text bytea
field returned an array identity.

TimestamptzCodec.decodeTextAs(Time.class) went through TimestampUtils.toTime,
which anchors the time of day to 1970-01-01 in the literal's own offset - a
day later in UTC once the offset crosses midnight, where the binary branch
truncates the instant. Both print the same time, so only a comparison saw it.

+++ pgjdbc/src/main/java/org/postgresql/jdbc/PgSQLInputBinary.java
@@ -163,7 +163,7 @@ public final class PgSQLInputBinary extends PgSQLInput {
 
   @Override
   protected float decodeFloat() throws SQLException {
-    return (float) PrimitiveDecoders.asDouble(getCodec(), source, curOffset, curLength, getCurrentType(), ctx);
+    return PrimitiveDecoders.asFloat(getCodec(), source, curOffset, curLength, getCurrentType(), ctx);
   }
 
   @Override
